refactor(www): replace createGetSSRProps function
