+++ apps/www/pages/vorschau/index.js
@@ -1,7 +1,7 @@
 import { useRouter } from 'next/router'
 import Front from '../../components/Front'
-import createGetServerSideProps from '../../lib/helpers/createGetServerSideProps'
 import { FRONT_QUERY } from '../../components/Front/graphql/getFrontQuery.graphql'
+import { createGetServerSideProps } from '../../lib/apollo/helpers'
 
 const FRONT_PREVIEW_PATH = `/vorschau`
